Сделано задание 25 урока: Переход на 1 активити с плейсхолдером для фрагментов.
Реализованы 2 фрагмента:
1. RecyclerView со списком фильмов. Карточка каждого фильма реализована с помощью двух вложенных LinearLayout, ConstraintLayout в данном случае глючил. Добавлена реализация DiffUtils.
2. Фрагмент с деталями фильма. Реализована с помощью CoordinatorView. Добавлен снекбар на FAB.

В активити добавлены тосты на кнопки меню.

Реализован выход из приложения (сворачивание) по двойному нажатию назад с учётом стека фрагментов.
