+++ app/src/main/java/film/search/filmsearch/FilmViewHolder.kt
@@ -3,6 +3,7 @@ package film.search.filmsearch
 import androidx.recyclerview.widget.RecyclerView
 import film.search.filmssearch.databinding.FilmItemBinding
 
+// Holder for Recycler view item
 class FilmViewHolder(val binding: FilmItemBinding) : RecyclerView.ViewHolder(binding.root) {
 
 }
