+++ app/src/main/res/drawable/screen_background.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
-    <gradient android:startColor="@color/colorBottomMenu"
+    <gradient android:startColor="@color/colorPrimaryDark"
         android:endColor="@color/color_light_blue"
-        android:angle="225"/>
+        android:angle="240"/>
 </shape>
