+++ app/src/main/res/values-ru/strings.xml
@@ -41,4 +41,5 @@
     <string name="walle_description">Робот ВАЛЛ·И из года в год прилежно трудится на опустевшей Земле, очищая нашу планету от гор мусора, которые оставили после себя улетевшие в космос люди. Он и не представляет, что совсем скоро произойдут невероятные события, благодаря которым он встретит друзей, поднимется к звездам и даже сумеет изменить к лучшему своих бывших хозяев, совсем позабывших родную Землю.</string>
     <string name="share">Поделиться</string>
     <string name="send_to_friend">Послать фильм %s другу?</string>
+    <string name="double_tap_toast">Нажмите два раза для выхода</string>
 </resources>
