+++ app/src/main/res/values/themes.xml
@@ -1,8 +1,11 @@
 <resources xmlns:tools="http://schemas.android.com/tools">
     <!-- Base application theme. -->
     <style name="Base.Theme.FilmSearch" parent="Theme.MaterialComponents.Light.NoActionBar">
+        <item name="colorPrimaryDark">#FF2060A0</item>
 
     </style>
+
+    <style name="AppTheme" parent="Theme.MaterialComponents.Light.NoActionBar" />
 
     <style name="Theme.FilmSearch" parent="Base.Theme.FilmSearch" />
 </resources>
